fetcher/extractor/requirements.txt
imgaug
numpy
pycocotools
mrcnn
tensorflow==1.14
keras==2.0.8
IPython